+++ order-service/order-domain/order-application-service/src/main/java/com/food/ordering/system/order/service/domain/OrderCreateCommandHandler.java
@@ -2,89 +2,41 @@ package com.food.ordering.system.order.service.domain;
 
 
 import com.food.ordering.system.domain.OrderDomainService;
-import com.food.ordering.system.domain.entity.Customer;
-import com.food.ordering.system.domain.entity.Order;
-import com.food.ordering.system.domain.entity.Restaurant;
 import com.food.ordering.system.domain.event.OrderCreatedEvent;
-import com.food.ordering.system.domain.exception.DomainException;
-import com.food.ordering.system.domain.exception.OrderDomainException;
 import com.food.ordering.system.order.service.domain.dto.create.CreateOrderCommand;
 import com.food.ordering.system.order.service.domain.dto.create.CreateOrderResponse;
 import com.food.ordering.system.order.service.domain.mapper.OrderDataMapper;
-import com.food.ordering.system.order.service.domain.ports.output.repository.CustomerRepository;
-import com.food.ordering.system.order.service.domain.ports.output.repository.OrderRepository;
-import com.food.ordering.system.order.service.domain.ports.output.repository.RestaurantRepository;
+import com.food.ordering.system.order.service.domain.ports.output.message.publisher.payment.OrderCreatedPaymentRequestMessagePublisher;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
-import java.util.UUID;
 
 @Slf4j
 @Component
 public class OrderCreateCommandHandler {
 
-    private final OrderDomainService orderDomainService;
-
-    private final OrderRepository orderRepository;
-
-    private final CustomerRepository customerRepository;
-    private final RestaurantRepository restaurantRepository;
-
+    private final OrderCreateHelper orderCreateHelper;
     private final OrderDataMapper orderDataMapper;
 
+    private final OrderCreatedPaymentRequestMessagePublisher orderCreatedPaymentRequestMessagePublisher;
 
     public OrderCreateCommandHandler(OrderDomainService orderDomainService,
-                                     OrderRepository orderRepository,
-                                     CustomerRepository customerRepository,
-                                     RestaurantRepository restaurantRepository,
-                                     OrderDataMapper orderDataMapper) {
-        this.orderDomainService = orderDomainService;
-        this.orderRepository = orderRepository;
-        this.customerRepository = customerRepository;
-        this.restaurantRepository = restaurantRepository;
+                                     OrderCreateHelper orderCreateHelper,
+                                     OrderDataMapper orderDataMapper,
+                                     OrderCreatedPaymentRequestMessagePublisher orderCreatedPaymentRequestMessagePublisher) {
+        this.orderCreateHelper = orderCreateHelper;
         this.orderDataMapper = orderDataMapper;
+        this.orderCreatedPaymentRequestMessagePublisher = orderCreatedPaymentRequestMessagePublisher;
     }
 
-    @Transactional
     public CreateOrderResponse createOrder(CreateOrderCommand createOrderCommand) {
 
-        checkCustomer(createOrderCommand.getCustomerId());
-        Restaurant restaurant = checkRestaurant(createOrderCommand);
-        Order order = orderDataMapper.createOrderCommandToOrder(createOrderCommand);
-        OrderCreatedEvent orderCreatedEvent = orderDomainService.validateAndInitiateOrder(order,restaurant);
-        Order orderResult = saveOrder(order);
-        log.info("Order created successfully with id {}", orderResult.getId().getValue());
-        return orderDataMapper.orderToCreateOrderResponse(orderResult);
+        OrderCreatedEvent orderCreatedEvent = orderCreateHelper.persistOrder(createOrderCommand);
+        log.info("Order created successfully with id {}", orderCreatedEvent.getOrder().getId().getValue());
+        orderCreatedPaymentRequestMessagePublisher.publish(orderCreatedEvent);
+        return orderDataMapper.orderToCreateOrderResponse(orderCreatedEvent.getOrder());
     }
 
-    private Restaurant checkRestaurant(CreateOrderCommand createOrderCommand) {
-       Restaurant restaurant = orderDataMapper.createOrderCommandToRestaurant(createOrderCommand);
-       Optional<Restaurant>  optionalRestaurant = restaurantRepository.findRestaurantInformation(restaurant);
-        if(optionalRestaurant.isEmpty()){
-            log.warn("Could not find Restaurant with  id: {}",createOrderCommand.getRestaurantId() );
-            throw new OrderDomainException("Could not find Restaurant with  id: " + createOrderCommand.getRestaurantId());
-        }
-        return optionalRestaurant.get();
-    }
 
-    private void checkCustomer(UUID customerId) {
-        Optional<Customer> customer = customerRepository.findCustomer(customerId);
-
-        if(customer.isEmpty()){
-            log.warn("Could not find Customer with customer id: {}", customerId);
-            throw new OrderDomainException("Could not find Customer with customer id: "+ customerId.toString());
-        }
-    }
-
-    private Order saveOrder(Order order){
-        Order orderResult =  orderRepository.save(order);
-        if(orderResult == null){
-            throw new DomainException("Could not save Order !");
-        }
-        log.info("Order is saved with id {} ", orderResult.getId().getValue());
-        return orderResult;
-    }
 
 }
